refactored NULL to C++11 nullptr
